requirements.txt
streamlit==1.28.1
pandas==2.1.3
plotly==5.17.0
openpyxl==3.1.2
mysql-connector-python==8.2.0
requests==2.31.0
matplotlib==3.7.2 